added spectral.md file. updated requirements.txt, modules.md,and tutorial.ipynb to include the spectral module and the new notebook example. Fixed some minor bugs in: -preprocessing (build_data_autoregressive indices made as int), -auxiliary(padgamma function expecting lags not only as list), -graphics (rescale signal when plotting together with vpath, fix ticks in the state means and covariance plot), -spectral (no indices resizing for one-channel data)

+++ glhmm/preproc.py
@@ -361,7 +361,7 @@ def build_data_autoregressive(data,indices,autoregressive_order=1,
     
     X = np.zeros((T - N*autoregressive_order,p*autoregressive_order))
     Y = np.zeros((T - N*autoregressive_order,p))
-    indices_new = np.zeros((N,2))
+    indices_new = np.zeros((N,2)).astype(int)
 
     for j in range(N):
         ind_1 = np.arange(indices[j,0]+autoregressive_order,indices[j,1],dtype=np.int64)
